requirements.txt
redshift_connector
pandas
numpy
python-dotenv
